+++ lib/protobuf/generators/enum_generator.rb
@@ -4,18 +4,29 @@ module Protobuf
   module Generators
     class EnumGenerator < Base
 
+      def allow_alias?
+        descriptor.options.try(:allow_alias!) { false }
+      end
+
       def compile
         run_once(:compile) do
           tags = []
 
           print_class(descriptor.name, :enum) do
+            if allow_alias?
+              puts "set_option :allow_alias"
+              puts
+            end
+
             descriptor.value.each do |enum_value_descriptor|
               tags << enum_value_descriptor.number
               puts build_value(enum_value_descriptor)
             end
           end
 
-          self.class.validate_tags(fully_qualified_type_namespace, tags)
+          unless allow_alias?
+            self.class.validate_tags(fully_qualified_type_namespace, tags)
+          end
         end
       end